[refactor] 코드 일부 정리
