+++ src/main/java/com/prgrms/rg/RgApplication.java
@@ -6,7 +6,7 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
-import com.prgrms.rg.infrastructure.cloud.message.CriticalMessageSender;
+import com.prgrms.rg.infrastructure.cloud.CriticalMessageSender;
 
 @EnableJpaAuditing
 @SpringBootApplication
